requirements.txt
flask
waitress
python-docx